% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_ewd.R
\name{agreguj_ewd}
\alias{agreguj_ewd}
\title{Obliczanie "wieloletnich" wskaznikow EWD poprzez agregacje wskaznikow "jednorocznych"}
\usage{
agreguj_ewd(
  wskazniki,
  lata,
  pliki = NULL,
  zapisz = TRUE,
  wskaznikiWR = length(wskazniki) == 1,
  usunPrzyrostki = "_.*$",
  traktujLataJakoNiezalezne = TRUE,
  src = NULL
)
}
\arguments{
\item{wskazniki}{wektor ciąg znaków z nazwami wskaźników, które mają zostać
zagregowane lub wyrażeniami regularnymi, do których nazwy takich wskaźników
mają pasować - por. argument \code{wskaznikiWR}}

\item{lata}{wektor liczb naturalnych - lata, które mają zostać
zagregowane}

\item{pliki}{opcjonalnie wektor ciągów znaków z nazwami plików (zapisanych
na dysku np. w wyniku użycia \code{oblicz_ewd_bk()}) zawierających wartości
wskaźników, z których mają one zostać wczytane zamiast z bazy danych}

\item{zapisz}{wartość logiczna - czy zapisać obliczone wartości zagregowanych
wskaźników EWD na dysku w formie pliku .RData?}

\item{wskaznikiWR}{wartość logiczna - czy traktować argument \code{wskazniki}jako
zawierający wyrażenia regularne (a nie nazwy wskaźników)? domyślnie \code{TRUE},
jeśli argument \code{wskaźniki} to pojedynczy ciąg znaków, a \code{FALSE}, jeśli jest
to wektor kilku ciągów znaków}

\item{usunPrzyrostki}{jeśli wartości wskaźników są pobierane z bazy (tj.
\code{pliki=NULL}), ciąg znaków z wyrażeniem regularnym opisującym ciąg znaków,
który powinien zostać usunięty z nazw wskaźników pobranych z bazy (por.
sekcja \emph{details} poniżej); jeśli z nazw wskaźników pobranych z bazy nie ma
być usuwany żaden ciąg znaków, jako wartość argumentu należy podać pusty
ciąg znaków (\code{""})}

\item{traktujLataJakoNiezalezne}{wartość logiczna - czy przy obliczaniu
(ko)wariancji traktować wskaźniki z poszczególnych lat jako niezależne od
siebie, czy jako powiązane zmienne? (obliczając empiryczne kowariancje
pomiędzy wskaźnikami w poszczególnych latach) - obecnie jedyna możliwa do
wyboru wartość to \code{TRUE}, ale argument ma przypominać, że można by
zaimplementować i alternatywne podejście}

\item{src}{połączenie z bazą danych IBE zwracane przez funkcję \code{\link[ZPD:polacz]{ZPD::polacz()}};
jeśli nie podane, podjęta zostanie próba automatycznego nawiązania połączenia
(poprzez wywoływanie funkcji \code{\link[ZPD:polacz]{ZPD::polacz()}} z domyślnymi argumentami)}
}
\value{
lista klasy \emph{listaWskaznikowEWD}
}
\description{
Oblicza \emph{wieloletnie} wskaźniki EWD poprzez agregację wskaźników
\emph{jednorocznych}.
}
\details{
Podając wartości argumentu \code{wskazniki} należy mieć na uwadze, czy wartości
wskaźników mają być pobrane z bazy danych, czy wczytane z zapisanych na dysku
plików (jeśli zostanie podany argument \code{pliki}) - nazwy wskaźników w bazie
oprócz ciągu znaków identyfikującego egzamin na wyjściu, typ szkoły
i przedmiot (obszar), np. \emph{mlb}, zawierają również przyrostki identyfikujące
wersję wskaźnika (w tym odróżniające wskaźniki \emph{jednoroczne} od
\emph{wieloletnich}!), podczas gdy w plikach zapisanych na dysku nazwy wskaźników
są tych przyrostków pozbawione (są one dodawane dopiero na etapie zapisywania
wartości wskaźników do bazy).
}
\seealso{
\code{\link[=oblicz_ewd_bk]{oblicz_ewd_bk()}}
}
