% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_wybor_czesci_egzaminu.R
\name{dodaj_wybor_czesci_egzaminu}
\alias{dodaj_wybor_czesci_egzaminu}
\title{Przetwarzanie danych z wynikami egzaminow}
\usage{
dodaj_wybor_czesci_egzaminu(dane, czesciEgzaminow)
}
\arguments{
\item{dane}{ramka danych z wynikami egzaminu}

\item{czesciEgzaminow}{ramka danych z charakterystykami (pseudo)kryteriów
oceny skali (występujących w wynikach przekazanych argumentem \code{dane}),
zwrócona przez [przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())]

[przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())]: R:przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())}
}
\value{
ramka danych przekazana argumentem \code{dane}, z dodanymi zmiennymi
opisującymi wybór (zdawanie) poszczególnych części egzaminu
}
\description{
Funkcja dodaje do danych z wynikami egzaminu zmienne (pseudokryteria)
opisujące wybór części egzaminu.
}
\details{
Zmienne (pseudokryteria) opisujące fakt przystąpienia (lub) nie do
poszczególnych części egzaminu umożliwiają modelowanie różnic w średnim
poziomie (ale już \strong{nie} w zróżnicowaniu) umiejętności pomiędzy uczniami
wybierającymi różne przedmioty (na różnych poziomach) w sytuacji, gdy
skalowanych jest jednocześnie wiele przedmiotów i modelowanie takich różnic
poprzez grupy w modelu wielogrupowym prowadziłoby do rozpatrywania zbyt
dużej liczby grup, z których niektóre były by bardzo małoliczne.

W ramach przyjętej procedury skalowania podejście to
\strong{nie jest wykorzystywane}, niemniej funkcja pozostaje włączona do pakietu.
}
