% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estymuj_pvreg.R
\name{estymuj_pvreg}
\alias{estymuj_pvreg}
\title{Obliczanie latentnych wskaznikow EWD}
\usage{
estymuj_pvreg(
  dane,
  parametry,
  nazwa,
  nPV,
  nWatkow,
  nadpisz,
  metoda = c("tylko pliki", "Python", "R")
)
}
\arguments{
\item{dane}{lista ramek danych zwrócona przez \code{\link[=przygotuj_dane_do_ewd_bk]{przygotuj_dane_do_ewd_bk()}}}

\item{parametry}{lista ramek danych zwrócona przez
\code{\link[=pobierz_parametry_egzaminow]{pobierz_parametry_egzaminow()}}, a następnie zmodyfikowana przez
\code{\link[=unormuj_parametry_egzaminow]{unormuj_parametry_egzaminow()}}}

\item{nazwa}{ciąg znaków - podstawa do tworzenia nazw plików z danymi dla
\emph{pvreg} i wynikami zapisywanymi przez \emph{pvreg}}

\item{nPV}{liczba (naturalna) - liczba PV z indywidualnymi oszacowaniami
umiejętności, które mają zostać wykorzystane przez \emph{pvreg}}

\item{nWatkow}{liczba (naturalna) - liczba równoległych wątków, w których
\emph{pvreg} będzie prowadził estymację (co do zasady powinna być dzielnikiem
\code{nPV})}

\item{nadpisz}{wartość logiczna - czy jeśli na dysku znajdują się już
zapisane wartości wskaźników EWD obliczone przez \emph{pvreg}, to nadpisać je
(\code{nadpisz=TRUE}), czy wczytać z nich wyniki bez uruchamiania \emph{pvreg}
(\code{nadpisz=FALSE}); wartość argumentu nie ma znaczenia, jeśli argument
\code{metoda="tylko pliki"}}

\item{metoda}{ciąg znaków (można podawać skrócony) wskazujący, w jaki sposób
funkcja ma zadziałać:
\itemize{
\item \code{"tylko pliki"} - funkcja przygotowuje pliki z danymi (wynikami
i parametrami) oraz pliki sterujące dla \emph{pvreg}, ale \strong{nie} uruchamia go
w celu obliczenia wartości wskaźników EWD; użytkownik może później
uruchomić \emph{pvreg} samodzielnie z konsoli systemowej;
\item \code{"Python"} - funkcja zapisze na dysku potrzebne pliki, a następnie uruchomi
\emph{pvreg} w celu obliczenia wskaźników, wczyta zapisane przez \emph{pvreg}
wartości wskaźników i przetworzy je do formy, w której będą mogły zostać
łatwo zapisane do bazy; \strong{jeśli funkcja znajdzie na dysku pliki z wynikami
działania \emph{pvreg}} i argument \code{nadpisz=FALSE}, to (dla danej skali) nie
będzie go uruchamiać, tylko wczyta wyniki z tych plików;
\item \code{"R"} - podjęta została próba przepisania \emph{pvreg} do R, która jednak
nie zakończyła się sukcesem - w związku z tym opcji tej w praktyce nie
można użyć (funkcja zakończy się błędem), ale jej istnienie przypomina
o tym wysiłku
}}
}
\value{
lista dwóch ramek danych z elementami \emph{ewd} i \emph{pv}
}
\description{
Przeprowadza estymację latentnych wskaźników EWD z użyciem pakietu \emph{pvreg}
Pythona. Odpowiada też za zapisanie na dysku danych dla \emph{pvreg} i wczytanie
wyników zapisanych przez \emph{pvreg} na dysku.
}
\seealso{
\code{\link[=zapisz_dane_dla_pvreg]{zapisz_dane_dla_pvreg()}}, \code{\link[=oblicz_ewd_bk]{oblicz_ewd_bk()}}
}
