% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_empiryczne_warstwice.R
\name{kwantyl_wazony}
\alias{kwantyl_wazony}
\title{Przetwarzanie i zapis wynikow obliczania latentnych wskaznikow EWD}
\usage{
kwantyl_wazony(x, wagi, probs = seq(0, 1, 0.25), na.rm = FALSE)
}
\arguments{
\item{x}{wektor liczb, dla którego mają zostać obliczone kwantyle}

\item{wagi}{wektor liczb nieujemnych z wagami, jakie powinny zostać
przypisane kolejnym elementom wektora \code{x}}

\item{probs}{wektor liczb z przedziału [0, 1] - kwantyle, które powinny
zostać obliczone}

\item{na.rm}{wartość logiczna - czy przy obliczaniu kwantyli ignorować ew.
braki danych w \code{x}}
}
\value{
wektor liczb
}
\description{
Oblicza kwantyle ważone.
}
\details{
Funkcja pierwotnie znajdowała się w pakiecie \emph{EWDogolny}.
}
\seealso{
\code{\link[=oblicz_empiryczne_warstwice]{oblicz_empiryczne_warstwice()}}
}
