% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_empiryczne_warstwice.R
\name{oblicz_empiryczne_warstwice}
\alias{oblicz_empiryczne_warstwice}
\title{Przetwarzanie i zapis wynikow obliczania latentnych wskaznikow EWD}
\usage{
oblicz_empiryczne_warstwice(wyniki, ewd, liczbaUczniow, pr = c(0.5, 0.9))
}
\arguments{
\item{wyniki}{wektor liczb - oszacowania średnich wyników egzaminu na
wyjściu poszczególnych szkół}

\item{ewd}{wektor liczb - oszacowania EWD poszczególnych szkół}

\item{liczbaUczniow}{wektor liczb (dodatnich) - liczba uczniów
uwzględnionych przy obliczaniu danego wskaźnika EWD w poszczególnych szkołach}

\item{pr}{wektor liczb z przedziału (0, 1) - częstości, dla których mają
zostać wyznaczone odcinające je w empirycznych danych wartości dystrybuanty
rozkładu chi-kwadrat o dwóch stopniach swobody}
}
\value{
wektor liczb (prawdopodobieństw) o długości równej długości
argumentu \code{pr}
}
\description{
Oblicza parametry warstwic EWD w populacji szkół ważonych liczbą uczniów,
analizując empiryczny rozkład wartości wskaźników w dwuwymiarowej przestrzeni
(wyniki_egzaminu_na_wyjściu x EWD).
}
\seealso{
\code{\link[=kwantyl_wazony]{kwantyl_wazony()}}, \code{\link[=przygotuj_ewd_do_zapisu]{przygotuj_ewd_do_zapisu()}}
}
