% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_ewd_bk.R
\name{oblicz_ewd_bk}
\alias{oblicz_ewd_bk}
\title{Obliczanie latentnych wskaznikow EWD szkol maturalnych}
\usage{
oblicz_ewd_bk(
  rokEWD,
  skaleMatura = paste0("^ewd;m_[^;]+;", rokEWD),
  skaleWe = paste0("^ewd;g[hm](LO;", rokEWD - 3L, rokEWD - 4L, "|T;", rokEWD - 4L, rokEWD
    - 5L, ")"),
  zapisz = metoda != "tylko pliki",
  katalogSurowe = "../../skalowanie/dane surowe",
  skalowanie = -1L,
  nadpisz = FALSE,
  metoda = c("tylko pliki", "Python", "R"),
  minLUcznSzk = 1L,
  nPV = 5L,
  nWatkow = 5L,
  czyEgzaminOsmoklasisty = NA,
  src = NULL
)
}
\description{
Obliczanie latentnych wskaznikow EWD szkol maturalnych
}
