% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrob_wyniki_uirt.R
\name{obrob_wyniki_uirt}
\alias{obrob_wyniki_uirt}
\title{Skalowanie wynikow egzaminow w Stacie przy pomocy pakietu uirt}
\usage{
obrob_wyniki_uirt(parametry, dane)
}
\arguments{
\item{parametry}{ramka danych z wyestymowanymi parametrami modelu
(wczytanymi wprost z pliku CSV zapisanego przez Statę)}

\item{dane}{ramka danych z wynikami egzaminu do wyskalowania - co do zasady
zwrócona przez \code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}}
}
\value{
lista z wynikami skalowania zawierająca następujące elementy:
\emph{grupy} (ramka danych), \emph{parametry} (ramka danych), \emph{kowariancje}
(ramka danych)
}
\description{
Przetwarza wyniki skalowania z formy zapisanej ze Staty do kompatybilnej
ze strukturami danych, w jakich są one zapisywane w bazie danych.
}
\seealso{
\code{\link[=skaluj_uirt]{skaluj_uirt()}}
}
