% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okresl_grupe.R
\name{okresl_grupe}
\alias{okresl_grupe}
\title{Przetwarzanie danych z wynikami egzaminow}
\usage{
okresl_grupe(
  dane,
  rokWy,
  katalogSurowe,
  typSzkoly = NULL,
  rokWe = NULL,
  echo = TRUE
)
}
\arguments{
\item{dane}{ramka danych z wynikami egzaminu}

\item{rokWy}{liczba (całkowita) - rok matury (i jednocześnie rok definiujący
jednoroczny wskaźnik EWD, na potrzeby obliczania którego ma być
przeprowadzone skalowanie, do którego przygotowywane są dane)}

\item{katalogSurowe}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów oraz \emph{danymi kontekstowymi}, pobranymi
przy pomocy funkcji \link[EWDdane:pobierz_wyniki_surowe]{EWDdane::pobierz_wyniki_surowe}}

\item{typSzkoly}{\code{"LO"}, \code{"T"} albo \code{NULL} - jeśli przetwarzane dane dotyczą
egzaminu na wejściu, typ szkoły, który obejmują dane - niezbędny, aby
określić, \emph{dane kontekstowe} z którego roku opisują uczniów o typowej
długości toku kształcenia, a z którego roku opisują uczniów o toku
kształcenia wydłużonym o rok (jeśli przetwarzane dane dotyczą matury,
należy podać \code{NULL})}

\item{rokWe}{liczba (całkowita) albo \code{NULL} - jeśli przetwarzane dane dotyczą
egzaminu na wejściu, typ szkoły, który obejmują dane - niezbędny, aby
określić, \emph{dane kontekstowe} z którego roku opisują uczniów o typowej
długości toku kształcenia, a z którego roku opisują uczniów o toku
kształcenia wydłużonym o rok (jeśli przetwarzane dane dotyczą matury,
należy podać \code{NULL})}

\item{echo}{wartość logiczna - czy wyświetlić na konsoli tabelę z rozkładem
grup}
}
\value{
ramka danych przekazana argumentem \code{dane}, z dodaną zmienną \emph{grupa}
opisującą przypisanie uczniów do grup wyróżnianych w skalowaniu
}
\description{
Dodaje do danych zmienną opisującą przypisanie uczniów do grup wyróżnianych
w procedurze skalowania wyników matury lub egzaminu na wejściu.
}
\details{
W odniesieniu do matury wyróżniane grupy to:
\itemize{
\item dla przedmiotów obowiązkowych: \emph{LO PP i PR}, \emph{LO PP}, \emph{T PP i PR}, \emph{T PP}
i \emph{inne};
\item dla przedmiotów nieobowiązkowych: \emph{LO PR}, \emph{T PR} i \emph{inne}.
W odniesieniu do egzaminów na wejściu (egzamin ósmoklasisty lub gimnazjalny):
\item dla LO: \emph{LO}, \emph{inne}, \emph{LOw}, \emph{innew} (dwie ostatnie są tworzone przez
uczniów zdających egzamin w roku o jeden wcześniejszym, niż wynikałoby to
z długości toku kształcenia w LO),
\item dla techników: \emph{T}, \emph{inne}, \emph{Tw}, \emph{innew} (dwie ostatnie są tworzone przez
uczniów zdających egzamin w roku o jeden wcześniejszym, niż wynikałoby to
z długości toku kształcenia w technikum).#'
}
}
\seealso{
\code{\link[=wyswietl_rozklad_grup]{wyswietl_rozklad_grup()}}, \code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}
}
