% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okresl_liczbe_uczniow_w_szkolach.R
\name{okresl_liczbe_uczniow_w_szkolach}
\alias{okresl_liczbe_uczniow_w_szkolach}
\title{Przygotowywanie danych do obliczenia wskaznikow EWD}
\usage{
okresl_liczbe_uczniow_w_szkolach(rokEWD, katalogSurowe, src)
}
\arguments{
\item{rokEWD}{rok egzaminu maturalnego (identyfikujący rocznik absolwentów)}

\item{katalogSurowe}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\link[EWDdane:pobierz_wyniki_surowe]{EWDdane::pobierz_wyniki_surowe}}

\item{src}{połączenie z bazą danych IBE zwracane przez funkcję \code{\link[ZPD:polacz]{ZPD::polacz()}};
jeśli nie podane, podjęta zostanie próba automatycznego nawiązania połączenia
(poprzez wywoływanie funkcji \code{\link[ZPD:polacz]{ZPD::polacz()}} z domyślnymi argumentami)}
}
\value{
ramka danych z kolumnami \emph{id_szkoly}, \emph{lu_wszyscy}, \emph{rok},
\emph{matura_miedzynarodowa}
}
\description{
Na podstawie wczytanych z dysku \emph{danych kontekstowych} dotyczących matury
określa liczbę wszystkich absolwentów poszczególnych szkół w danym roczniku
i dołącza do tego zestawienia informację z bazy szkół, czy w danej szkole
przeprowadzana jest matura międzynarodowa.
}
\seealso{
\code{\link[ZPD:pobierz_szkoly]{ZPD::pobierz_szkoly()}}, \code{\link[=oblicz_ewd_bk]{oblicz_ewd_bk()}}
}
