% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okresl_liczbe_zdajacych.R
\name{okresl_liczbe_zdajacych}
\alias{okresl_liczbe_zdajacych}
\title{Przetwarzanie i zapis wynikow obliczania latentnych wskaznikow EWD}
\usage{
okresl_liczbe_zdajacych(dane, rokEWD, luWszyscy, nazwaWskaznika)
}
\arguments{
\item{dane}{lista ramek danych zwrócona przez \code{\link[=przygotuj_dane_do_ewd_bk]{przygotuj_dane_do_ewd_bk()}}}

\item{rokEWD}{rok egzaminu maturalnego (identyfikujący rocznik absolwentów)}

\item{luWszyscy}{ramka danych zwrócona przez
\code{\link[=okresl_liczbe_uczniow_w_szkolach]{okresl_liczbe_uczniow_w_szkolach()}}}

\item{nazwaWskaznika}{ciąg znaków - nazwa obliczonego wskaźnika EWD}
}
\value{
ramka danych
}
\description{
Określa liczbę zdających poszczególne przedmioty na poszczególnych poziomach
w danych użytych do obliczenia wskaźników EWD.
}
\seealso{
\code{\link[=przygotuj_ewd_do_zapisu]{przygotuj_ewd_do_zapisu()}}
}
