% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okresl_poziom_egzaminu.R
\name{okresl_poziom_egzaminu}
\alias{okresl_poziom_egzaminu}
\title{Przetwarzanie danych z wynikami egzaminow}
\usage{
okresl_poziom_egzaminu(
  dane,
  czesciEgzaminow,
  zmienNazweGdyTylkoJednaCzesc = TRUE
)
}
\arguments{
\item{dane}{ramka danych z wynikami egzaminu}

\item{czesciEgzaminow}{ramka danych z charakterystykami (pseudo)kryteriów
oceny skali (występujących w wynikach przekazanych argumentem \code{dane}),
zwrócona przez [przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())]

[przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())]: R:przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())}

\item{zmienNazweGdyTylkoJednaCzesc}{wartość logiczna - p. sekcja \emph{details}}
}
\value{
ramka danych przekazana argumentem \code{dane}, z dodanymi zmiennymi
opisującymi poziom(y), na których poszczególni uczniowie zdawali poszczególne
przedmioty matury
}
\description{
W odniesieniu do wyników egzaminu maturalnego, tworzy zmienne opisujące
poziom(y) poszczególnych przedmiotów zdawane przez poszczególnych uczniów.
}
\details{
Jeśli przekazane dane obejmują kilka różnych przedmiotów (w sensie różnych
wartości kolumny \code{czesc_egzaminu} ramki danych przekazanej do funkcji
argumentem \code{czesciEgzaminow} - p. \link{przygotuj_mapowanie_kryteriow_na_czesci_egzaminu}),
a także zawsze wtedy, kiedy argument \code{zmienNazweGdyTylkoJednaCzesc=FALSE},
nazwy zmiennych opisujących poziomy, na których poszczególni uczniowie
zdawali poszczególne przedmioty brane są z połączenia wartości kolumn
\code{czesc_egzaminu} i \code{poziom} ramki danych przekazanej do funkcji argumentem
\code{czesciEgzaminow} Jednak jeśli przekazane dane obejmują tylko jeden przedmiot
i \code{zmienNazweGdyTylkoJednaCzesc=TRUE} (wartość domyślna), nazwa kolumny
opisującej poziom zdawanego przedmiotu zostanie ustawiona na \emph{poziom}.#'
}
\seealso{
\code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}
}
