% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_parametry_egzaminow.R
\name{pobierz_parametry_egzaminow}
\alias{pobierz_parametry_egzaminow}
\title{Przygotowywanie danych do obliczenia wskaznikow EWD}
\usage{
pobierz_parametry_egzaminow(skale, src)
}
\arguments{
\item{skale}{ramka danych o jednym wierszu zawierająca kolumny
\emph{id_skali_matura}, \emph{skalowanie_matura}, \emph{id_skali_we}, \emph{skalowanie_we}
(pojedynczy wiersz połączonych ze sobą i przekształconych wywołań
\code{\link[=znajdz_skale]{znajdz_skale()}} dla skal matury i skal egzaminu na wejściu)}

\item{src}{połączenie z bazą danych IBE zwracane przez funkcję \code{\link[ZPD:polacz]{ZPD::polacz()}};
jeśli nie podane, podjęta zostanie próba automatycznego nawiązania połączenia
(poprzez wywoływanie funkcji \code{\link[ZPD:polacz]{ZPD::polacz()}} z domyślnymi argumentami)}
}
\value{
lista ramek danych z elementami: \emph{parametryMatura},
\emph{kowariancjeMatura}, \emph{tematyLaureatowMatura}, \emph{parametryWejscie},
\emph{kowariancjeWejscie}, \emph{tematyLaureatowWejscie}
}
\description{
Dla podanych skal pobiera z bazy zestawienie wartości parametrów modeli
skalowania i ich kowariancji.
}
\seealso{
\code{\link[ZPD:pobierz_parametry]{ZPD::pobierz_parametry()}},
\code{\link[ZPD:pobierz_skalowania_elementy_kowariancje]{ZPD::pobierz_skalowania_elementy_kowariancje()}},
\code{\link[=wybierz_tematy_dla_laureatow]{wybierz_tematy_dla_laureatow()}} \code{\link[=oblicz_ewd_bk]{oblicz_ewd_bk()}}
}
