% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wartosci_wskaznikow_ewd.R
\name{pobierz_wartosci_wskaznikow_ewd}
\alias{pobierz_wartosci_wskaznikow_ewd}
\title{Obliczanie "wieloletnich" wskaznikow EWD poprzez agregacje wskaznikow "jednorocznych"}
\usage{
pobierz_wartosci_wskaznikow_ewd(
  wskazniki,
  lata,
  wskaznikiWR,
  usunPrzyrostki,
  src
)
}
\arguments{
\item{wskazniki}{wektor ciąg znaków z nazwami wskaźników, które mają zostać
zagregowane lub wyrażeniami regularnymi, do których nazwy takich wskaźników
mają pasować - por. argument \code{wskaznikiWR}}

\item{lata}{wektor liczb naturalnych - lata, które mają zostać
zagregowane}

\item{wskaznikiWR}{wartość logiczna - czy traktować argument \code{wskazniki}jako
zawierający wyrażenia regularne (a nie nazwy wskaźników)? domyślnie \code{TRUE},
jeśli argument \code{wskaźniki} to pojedynczy ciąg znaków, a \code{FALSE}, jeśli jest
to wektor kilku ciągów znaków}

\item{usunPrzyrostki}{jeśli wartości wskaźników są pobierane z bazy (tj.
\code{pliki=NULL}), ciąg znaków z wyrażeniem regularnym opisującym ciąg znaków,
który powinien zostać usunięty z nazw wskaźników pobranych z bazy (por.
sekcja \emph{details} poniżej); jeśli z nazw wskaźników pobranych z bazy nie ma
być usuwany żaden ciąg znaków, jako wartość argumentu należy podać pusty
ciąg znaków (\code{""})}

\item{src}{połączenie z bazą danych IBE zwracane przez funkcję \code{\link[ZPD:polacz]{ZPD::polacz()}};
jeśli nie podane, podjęta zostanie próba automatycznego nawiązania połączenia
(poprzez wywoływanie funkcji \code{\link[ZPD:polacz]{ZPD::polacz()}} z domyślnymi argumentami)}
}
\value{
lista ramek danych z elementami \emph{ewd}, \emph{wskazniki},
\emph{wskaznikiSkalowania} i \emph{liczbaZdajacych}
}
\description{
Pobiera wartości wskaźników EWD (oraz powiązane z nimi informacje) z bazy
danych
}
\seealso{
\code{\link[=agreguj_ewd]{agreguj_ewd()}}
}
