% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podmien_wyniki_laureatow.R
\name{podmien_wyniki_laureatow}
\alias{podmien_wyniki_laureatow}
\title{Przygotowywanie danych do obliczenia wskaznikow EWD}
\usage{
podmien_wyniki_laureatow(dane, tematyLaureatow, zmienneTematy)
}
\arguments{
\item{dane}{ramka danych zawierająca wyniki egzaminu (oraz informacje
o (pseudo)kryteriach oceny jako atrybut \emph{czesciEgzaminu}), zwrócona przez
\code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}}

\item{tematyLaureatow}{ramka danych opisująca, jakie tematy rozprawki należy
przypisać laureatom w poszczególnych częściach egzaminu, zwrócona przez
\code{\link[=wybierz_tematy_dla_laureatow]{wybierz_tematy_dla_laureatow()}}}

\item{zmienneTematy}{wektor ciągów znaków zawierający nazwy zmiennych
opisujących wybór tematów}
}
\value{
ramka danych przekazana argumentem \code{dane}, w której odpowiednio
zmieniono laureatom wartości zmiennych opisujących wybór tematu rozprawki
i punktację (pseudo)kryteriów rozprawki
}
\description{
Zmienia wartości zmiennych opisujących wybór tematów rozprawki i punktację
(pseudo)kryteriów rozprawki laureatom poszczególnych części egzaminu.
}
\details{
Jeśli dla danej części egzaminu (w której były tematy) nie ma wskazanej nazwy
zmiennej kodującej wybór tematu, który należy przypisać laureatom, to
przypisuje im maksymalne wyniki z najczęściej wybieranego tematu spośród tych,
dla których w modelu skalowania nie ma zmiennych opisujących ich wybór.
}
\seealso{
\code{\link[=przygotuj_dane_do_ewd_bk]{przygotuj_dane_do_ewd_bk()}},
}
