% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane_do_ewd_bk.R
\name{przygotuj_dane_do_ewd_bk}
\alias{przygotuj_dane_do_ewd_bk}
\title{Przygotowywanie danych do obliczenia wskaznikow EWD}
\usage{
przygotuj_dane_do_ewd_bk(
  skale,
  rokEWD,
  typSzkoly,
  tematyLaureatowMatura,
  parametryMatura,
  tematyLaureatowWejscie,
  parametryWejscie,
  katalogSurowe,
  minLUcznSzk,
  src
)
}
\arguments{
\item{skale}{ramka danych o jednym wierszu zawierająca kolumny
\emph{id_skali_matura}, \emph{skalowanie_matura}, \emph{id_skali_we}, \emph{skalowanie_we}
(pojedynczy wiersz połączonych ze sobą i przekształconych wywołań
\code{\link[=znajdz_skale]{znajdz_skale()}} dla skal matury i skal egzaminu na wejściu)}

\item{rokEWD}{rok egzaminu maturalnego (identyfikujący rocznik absolwentów)}

\item{typSzkoly}{ciąg znaków - typ szkoły (co do zasady \code{"LO"} lub \code{"T"}) -
wykorzystywany do odfiltrowania parametrów modeli skalowania opisujących
rozkłady umiejętności w ramach grup dotyczących tylko tych grup, które
obejmują uczniów szkół danego typu}

\item{tematyLaureatowMatura}{ramka danych zwrócona przez
\code{\link[=wybierz_tematy_dla_laureatow]{wybierz_tematy_dla_laureatow()}}}

\item{parametryMatura}{ramka danych z parametrami modelu skalowania matury
(p. \link{pobierz_parametry_egzaminow})}

\item{tematyLaureatowWejscie}{ramka danych zwrócona przez
\code{\link[=wybierz_tematy_dla_laureatow]{wybierz_tematy_dla_laureatow()}}}

\item{parametryWejscie}{ramka danych z parametrami modelu skalowania egzaminu
na wejściu (p. \link{pobierz_parametry_egzaminow})}

\item{katalogSurowe}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\link[EWDdane:pobierz_wyniki_surowe]{EWDdane::pobierz_wyniki_surowe}}

\item{minLUcznSzk}{liczba (naturalna) - szkoły z liczbą uczniów (wchodzących
do obliczeń) poniżej zadanej liczby zostaną wykluczone z estymacji}

\item{src}{połączenie z bazą danych IBE zwracane przez funkcję \code{\link[ZPD:polacz]{ZPD::polacz()}};
jeśli nie podane, podjęta zostanie próba automatycznego nawiązania połączenia
(poprzez wywoływanie funkcji \code{\link[ZPD:polacz]{ZPD::polacz()}} z domyślnymi argumentami)}
}
\value{
lista ramek danych o następujących elementach: \emph{matura}, \emph{wejscie},
\emph{warunkujace}, \emph{cechySzkol}, z których trzy pierwsze zawierają dane do
wykorzystania w obliczeniach przez \emph{pvreg} a ostatnia informacje przydatne
w obróbce wyników i przygotowywaniu ich do zapisania do bazy danych
}
\description{
Przygotowuje zestaw danych do obliczania latentnych wskaźników EWD w formie
odpowiadającej interfejsowi wejściowemu \emph{pvreg}.
}
\seealso{
\code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}, \code{\link[=podmien_wyniki_laureatow]{podmien_wyniki_laureatow()}},
\code{\link[=wyswietl_rozklad_grup]{wyswietl_rozklad_grup()}}, \code{\link[=oblicz_ewd_bk]{oblicz_ewd_bk()}}, \code{\link[=unormuj_parametry_egzaminow]{unormuj_parametry_egzaminow()}}
}
