% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane_do_skalowania.R
\name{przygotuj_dane_do_skalowania}
\alias{przygotuj_dane_do_skalowania}
\title{Przetwarzanie danych z wynikami egzaminow}
\usage{
przygotuj_dane_do_skalowania(
  skala,
  rokEWD,
  katalogSurowe,
  dodajRokDoNazwTematow = FALSE,
  laureaci = c("usuń", "zostaw", "usuń rozprawki"),
  echo = TRUE,
  src = NULL
)
}
\arguments{
\item{skala}{ramka danych o jednym wierszy zawierająca kolumny
\emph{id_skali}, \emph{rodzaj_egzaminu}, \emph{lata} i \emph{typ_szkoly} (co do zasady pojedynczy
wiersz nieco przekształconego efektu wywołania \code{\link[=znajdz_skale]{znajdz_skale()}})}

\item{rokEWD}{rok egzaminu maturalnego (identyfikujący rocznik absolwentów)}

\item{katalogSurowe}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\link[EWDdane:pobierz_wyniki_surowe]{EWDdane::pobierz_wyniki_surowe}}

\item{dodajRokDoNazwTematow}{wartość logiczna - czy do tworzonych
nazw zmiennych opisujących wybór tematów ma zostać dopisany rok
przeprowadzenia egzaminu? można podać wartość \code{NA}, która oznacza, że rok
zostanie dodany tylko w sytuacji, gdy kryteria danej skali pochodzą
z egzaminów przeprowadzonych w kilku różnych latach}

\item{laureaci}{\itemize{
\item ciąg znaków opisujący, w jaki sposób postąpić z wynikami
laureatów
}}

\item{echo}{wartość logiczna - czy wyświetlić na konsoli tabele z rozkładami
wyboru tematów wypracowania i przypisania zdających do grup}

\item{src}{połączenie z bazą danych IBE zwracane przez funkcję \code{\link[ZPD:polacz]{ZPD::polacz()}};
jeśli nie podane, podjęta zostanie próba automatycznego nawiązania połączenia
(poprzez wywoływanie funkcji \code{\link[ZPD:polacz]{ZPD::polacz()}} z domyślnymi argumentami)}
}
\value{
ramka danych z wynikami egzaminu oraz dodatkowymi zmiennymi:
ew. \emph{pseudozadaniami} opisującymi wybór tematów wypracowania oraz zmienną
opisującą przypisanie zdających do grup
}
\description{
Funkcja przygotowuje dane do skalowania, wywołując kolejno bardziej
specjalizowane funkcje odpowiedzialne za obsłużenie laureatów, dodanie
zmiennych opisujących wybór tematów oraz przypisanie zdających do grup.
}
\seealso{
\code{\link[=wczytaj_wyniki_surowe]{wczytaj_wyniki_surowe()}},
\code{\link[=przygotuj_mapowanie_kryteriow_na_czesci_egzaminu]{przygotuj_mapowanie_kryteriow_na_czesci_egzaminu()}},
\code{\link[=usun_wyniki_laureatow]{usun_wyniki_laureatow()}}, \code{\link[=dodaj_wybory_tematow]{dodaj_wybory_tematow()}},
\code{\link[=okresl_poziom_egzaminu]{okresl_poziom_egzaminu()}}, \code{\link[=okresl_grupe]{okresl_grupe()}}, \code{\link[=skaluj_matura_bk]{skaluj_matura_bk()}},
\code{\link[=skaluj_we_bk]{skaluj_we_bk()}}, \code{\link[=przygotuj_dane_do_ewd_bk]{przygotuj_dane_do_ewd_bk()}}
}
