% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/przygotuj_mapowanie_kryteriow_na_czesci_egzaminu.R
\name{przygotuj_mapowanie_kryteriow_na_czesci_egzaminu}
\alias{przygotuj_mapowanie_kryteriow_na_czesci_egzaminu}
\title{Przetwarzanie danych z wynikami egzaminow}
\usage{
przygotuj_mapowanie_kryteriow_na_czesci_egzaminu(
  idSkali,
  dodajRokDoNazwTematow = FALSE,
  src = NULL
)
}
\arguments{
\item{idSkali}{liczba (całkowita) - id skali w bazie danych}

\item{dodajRokDoNazwTematow}{wartość logiczna - czy do tworzonych
nazw zmiennych opisujących wybór tematów ma zostać dopisany rok
przeprowadzenia egzaminu? można podać wartość \code{NA}, która oznacza, że rok
zostanie dodany tylko w sytuacji, gdy kryteria danej skali pochodzą
z egzaminów przeprowadzonych w kilku różnych latach}

\item{src}{połączenie z bazą danych IBE zwracane przez funkcję \code{\link[ZPD:polacz]{ZPD::polacz()}};
jeśli nie podane, podjęta zostanie próba automatycznego nawiązania połączenia
(poprzez wywoływanie funkcji \code{\link[ZPD:polacz]{ZPD::polacz()}} z domyślnymi argumentami)}
}
\value{
ramka danych, w której wiersze stanowią (pseudo)kryteria oceny
}
\description{
Funkcja na podstawie informacji z bazy danych zwraca (pseudo)kryteria oceny
skali o podanym id z dopisanymi informacjami o tym, z jakiej części egzaminu
i z jakiego roku pochodzą, prefiks związany z tą częścią egzaminu, numer
pytania, typ pytania oraz, w przypadku kryteriów oceny wypracowania, którego
tematu dane kryterium dotyczy (a także zmienną rozróżniającą części egzaminu
ze \emph{starej} i \emph{nowej} formuły matury, w sensie zmian z lat 2023/2034).
}
\seealso{
\code{\link[ZPD:pobierz_kryteria_oceny]{ZPD::pobierz_kryteria_oceny()}}, \code{\link[ZPD:pobierz_testy]{ZPD::pobierz_testy()}},
\code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}, \code{\link[=wybierz_tematy_dla_laureatow]{wybierz_tematy_dla_laureatow()}}
}
