% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_pv_do_zapisu.R
\name{przygotuj_pv_do_zapisu}
\alias{przygotuj_pv_do_zapisu}
\title{Przetwarzanie i zapis wynikow obliczania latentnych wskaznikow EWD}
\usage{
przygotuj_pv_do_zapisu(
  pv,
  skale,
  parametry,
  czesciEgzaminu,
  grupy,
  nazwaWskaznika,
  rokEWD
)
}
\arguments{
\item{pv}{ramka danych z wartościami PV wyestymowanymi przez \emph{pvreg}
(element \emph{PV} listy zwróconej przez \code{\link[=estymuj_pvreg]{estymuj_pvreg()}}, z wybranymi tylko
wierszami odnoszącymi się do danego egzaminu)}

\item{skale}{ramka danych o jednym wierszu zawierająca informacje o skali
i numerze skalowania, który powinien zostać przypisany wynikom estymacji
z wykorzystaniem \emph{pvreg}}

\item{parametry}{lista ramek danych z elementami \emph{parametry} i \emph{kowariancje}
z wartościami parametrów modelu skalowania danego egzaminu i ich kowariancji,
jakie zostały użyte w estymacji z wykorzystaniem \emph{pvreg} - co do zasady
wzięte z wyniku działania \code{\link[=pobierz_parametry_egzaminow]{pobierz_parametry_egzaminow()}}, przekształconego
następnie przy pomocy \code{\link[=unormuj_parametry_egzaminow]{unormuj_parametry_egzaminow()}}}

\item{czesciEgzaminu}{ramka danych - atrybut \emph{czesciEgzaminu} ramki danych
zawierającej wyniki danego egzaminu, będącej jednym z elementów listy
zwróconej przez wywołanie \code{\link[=przygotuj_dane_do_ewd_bk]{przygotuj_dane_do_ewd_bk()}}}

\item{grupy}{wektor ciągów znaków -  nazwy grup w modelu skalowania danego
egzaminu, które zostały objęte obliczaniem wskaźników EWD}

\item{nazwaWskaznika}{ciąg znaków - nazwa obliczonego wskaźnika EWD}

\item{rokEWD}{rok egzaminu maturalnego (identyfikujący rocznik absolwentów)}
}
\value{
lista klasy \emph{wynikiSkalowania}
}
\description{
Przetwarza zwrócone przez \emph{pvreg} oszacowania PV i parametrów modelu
skalowania danego egzaminu do formy odpowiadającej strukturom, w jakich
dane o wynikach skalowania zapisywane są w bazie danych.
}
\seealso{
\code{\link[=oblicz_ewd_bk]{oblicz_ewd_bk()}}, \code{\link[=przygotuj_ewd_do_zapisu]{przygotuj_ewd_do_zapisu()}}
}
