% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapisz_dane_dla_pvreg.R
\name{przygotuj_zbior_z_parametrami_dla_pvreg}
\alias{przygotuj_zbior_z_parametrami_dla_pvreg}
\title{Obliczanie latentnych wskaznikow EWD}
\usage{
przygotuj_zbior_z_parametrami_dla_pvreg(parametry, kowariancje, grupy)
}
\arguments{
\item{parametry}{ramka danych z parametrami modelu skalowania danego egzaminu}

\item{kowariancje}{ramka danych z kowariancjami parametrów modelu skalowania
danego egzaminu}

\item{grupy}{ramka danych z rozkładem liczebności grup w danych (i numerami
grup)}
}
\value{
ramka danych o wartościami parametrów modelu skalowania
i ich kowariancji, w strukturze odpowiadającej tej oczekiwanej przez \emph{pvreg}
}
\description{
Obliczanie latentnych wskaznikow EWD
}
\seealso{
\code{\link[=zapisz_dane_dla_pvreg]{zapisz_dane_dla_pvreg()}}
}
