% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_wybory_tematow.R
\name{przygotuj_zmienne_tematy_w_ramach_czesci_egzaminu}
\alias{przygotuj_zmienne_tematy_w_ramach_czesci_egzaminu}
\title{Przetwarzanie danych z wynikami egzaminow}
\usage{
przygotuj_zmienne_tematy_w_ramach_czesci_egzaminu(x, echo = TRUE)
}
\arguments{
\item{x}{ramka danych zawierające wyniki z pojedynczej części egzaminu
(z pojedynczego rok, w konkretnej - \emph{starej}/\emph{nowej} formule)}

\item{echo}{wartość logiczna - czy wyświetlić na konsoli tabelę z rozkładami
wyboru tematów wypracowania}
}
\value{
ramka danych przekazana argumentem \code{x} z dodaną (jeśli w danej części
egzaminu występuje rozprawka) zmienną opisującą wybór tematu rozprawki
(wypracowania)
}
\description{
\emph{Koń roboczy} \code{\link[=dodaj_wybory_tematow]{dodaj_wybory_tematow()}}.
}
