% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvreg.R
\name{pvreg}
\alias{pvreg}
\title{Proba sportowania programu Bartka Kondratka pod R}
\usage{
pvreg(
  dane,
  parametry,
  fixedEffects = setdiff(names(dane$warunkujace), c("id_obserwacji", "id_szkoly")),
  nPV = 5L,
  keepPV = TRUE,
  nJobs = NULL,
  nItBurn0 = 10L,
  nItBurn1 = 20L,
  nItFinal = 10L,
  echo = 1L,
  methodMultilevelPriors = c("simulateIndep", "simulateCond", "predRandResVar",
    "predRandAllVars", "predFixedAllVars"),
  rescalePriors = FALSE
)
}
\description{
Proba sportowania programu Bartka Kondratka pod R
}
