% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_az_dobrze.R
\name{skaluj_az_dobrze}
\alias{skaluj_az_dobrze}
\title{Skalowanie wynikow egzaminow w Stacie przy pomocy pakietu uirt}
\usage{
skaluj_az_dobrze(
  dane,
  zadania3pl,
  nazwa,
  nadpiszWynikiSkalowaniaNaDysku,
  maxNIter,
  usunUjemneDyskryminacje,
  uzyjRozkladowAPrioriDlaZadan3PL = list(a = c(1.5, 2), b = c(0, 2.5), c = c(2.5, 5.5)),
  kryteriaUsuwaniaZadan = c(absA = 0.1, absB = 10)
)
}
\arguments{
\item{dane}{ramka danych z wynikami egzaminu do wyskalowania - co do zasady
zwrócona przez \code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}}

\item{zadania3pl}{wektor ciągów znaków - nazwy (pseudo)kryteriów oceny
(zmiennych w ramce danych przekazanej argumentem \code{dane}), w odniesieniu do
których \emph{uirt} ma testować występowanie zgadywania (użycie modelu 3PL)}

\item{nazwa}{ciąg znaków - podstawa do tworzenia nazw plików, w których
zapisane zostaną zarówno dane i kod Staty, jak i wyniki skalowania
wyeksportowane ze Staty}

\item{nadpiszWynikiSkalowaniaNaDysku}{wartość logiczna - czy napotkawszy
zapisane na dysku wyniki skalowania zapisane przez \emph{uirt} funkcja powinna
przeprowadzić skalowanie i je nadpisać (\code{TRUE}), czy wczytać już istniejące,
nie przeprowadzając skalowania (\code{FALSE})?}

\item{maxNIter}{liczba (naturalna) - maksymalna liczba iteracji estymacji
\emph{uirt} w \emph{Stacie}}

\item{usunUjemneDyskryminacje}{wartość logiczna - czy \emph{uirt} powinien
automatycznie usuwać z modelu zadania o ujemnych dyskryminacjach?}

\item{uzyjRozkladowAPrioriDlaZadan3PL}{lista z wektorami dwuelementowymi
w formacie \code{c(wartoscOczekiwana, odchylenieStandardowe)} rozkładów \emph{a priori}
parametrów zadań 3PL, używanych przez \emph{uirt} w estymacji; lista może zawierać
elementy o nazwach \code{a}, \code{b} lub \code{c} - najlepiej wszystkie z nich; jeśli
\emph{uirt} ma nie używać rozkładów a priori dla zadań 3PL, należy jako wartość
tego argumentu podać \code{NULL} (ale co do zasady nie jest to dobre rozwiązanie);
co do zasady domyślną wartość tego argumentu należy uznać za rozsądną}

\item{kryteriaUsuwaniaZadan}{dwuelementowy wektor liczbowy, o nazwach
elementów \code{absA} i \code{absB} - zadania, których wyestymowane wartości
bezwzględne parametrów będą, odpowiednio: dyskryminacji (\code{absA}) - mniejsze,
niż podana wartość lub trudności (\code{absB}) - większe, niż podana wartość,
zostaną usunięte z estymacji; usuwanie zadań można efektywnie wyłączyć
poprzez podanie argumentu \code{kryteriaUsuwaniaZadan=c(absA=0, absB=Inf)}, ale
nie jest to dobry pomysł, gdyż używana przez \emph{uirt} metoda szacowania błędów
standardowych parametrów zadań nie działa dobrze dla zadań z dyskryminacją
bliską 0 (co zwykle idzie w parze z bardzo dużymi co do wartości bezwzględnej
oszacowaniami trudności); co do zasady domyślną wartość tego argumentu należy
uznać za rozsądną; uwaga, niezależnie od podanej wartości \code{absA} \strong{\emph{uirt}
zawsze będzie usuwał zadania o ujemnych wartościach dyskryminacji!}}
}
\value{
lista z wynikami skalowania zawierająca następujące elementy:
\emph{grupy} (ramka danych), \emph{parametry} (ramka danych), \emph{kowariancje}
(ramka danych), \emph{usunieteZadania} (wektor ciągów znaków, może mieć zerową
długość)
}
\description{
Skaluje wyniki egzaminu w pętli aż uda się spełnić podane argumentami
kryteria \emph{jakości} zadań (tzn. żadne zadanie nie zostanie usunięte na
podstawie wyestymowanych w \emph{uirt} wartości parametrów).
}
\details{
Zbyt niskie (bliskie zera; zadania o ujemnej dyskryminacji każemy \emph{uirt}-owi
usuwać automatycznie) wartości dyskryminacji sprawiają trudności przy
szacowaniu (ko)wariancji parametrów zadań metodą stosowaną przez \emph{uirt}
(choć skądinąd daje ona lepsze oszacowania tych (ko)wariancji, jeśli zadania
mają rozsądne własności pomiarowe). Funkcja pozwala dojść do takiego modelu,
który nie tylko się wyestymuje, ale również będzie miał kompletne
(i rozsądne) oszacowania (ko)wariancji parametrów modelu - jest to ważne,
gdyż również te (ko)wariancje będą potem wykorzystywane przy estymacji
wskaźników EWD przez \emph{pvreg}.
}
\seealso{
\code{\link[=skaluj_uirt]{skaluj_uirt()}}, \code{\link[=zwroc_zadania_niespelniajace_kryteriow]{zwroc_zadania_niespelniajace_kryteriow()}},
\code{\link[=skaluj_matura_bk]{skaluj_matura_bk()}}, \code{\link[=skaluj_we_bk]{skaluj_we_bk()}}
}
