% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_uirt.R
\name{skaluj_uirt}
\alias{skaluj_uirt}
\title{Skalowanie wynikow egzaminow w Stacie przy pomocy pakietu uirt}
\usage{
skaluj_uirt(
  dane,
  zadania3pl,
  nazwa,
  nadpiszWynikiSkalowaniaNaDysku = FALSE,
  maxNIter = 1000L,
  usunUjemneDyskryminacje = FALSE,
  uzyjRozkladowAPrioriDlaZadan3PL = list(a = c(1.5, 2), b = c(0, 2.5), c = c(2.5, 5.5))
)
}
\arguments{
\item{dane}{ramka danych z wynikami egzaminu do wyskalowania - co do zasady
zwrócona przez \code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}}

\item{zadania3pl}{wektor ciągów znaków - nazwy (pseudo)kryteriów oceny
(zmiennych w ramce danych przekazanej argumentem \code{dane}), w odniesieniu do
których \emph{uirt} ma testować występowanie zgadywania (użycie modelu 3PL)}

\item{nazwa}{ciąg znaków - podstawa do tworzenia nazw plików, w których
zapisane zostaną zarówno dane i kod Staty, jak i wyniki skalowania
wyeksportowane ze Staty}

\item{nadpiszWynikiSkalowaniaNaDysku}{wartość logiczna - czy napotkawszy
zapisane na dysku wyniki skalowania zapisane przez \emph{uirt} funkcja powinna
przeprowadzić skalowanie i je nadpisać (\code{TRUE}), czy wczytać już istniejące,
nie przeprowadzając skalowania (\code{FALSE})?}

\item{maxNIter}{liczba (naturalna) - maksymalna liczba iteracji estymacji
\emph{uirt} w \emph{Stacie}}

\item{usunUjemneDyskryminacje}{wartość logiczna - czy \emph{uirt} powinien
automatycznie usuwać z modelu zadania o ujemnych dyskryminacjach?}

\item{uzyjRozkladowAPrioriDlaZadan3PL}{lista z wektorami dwuelementowymi
w formacie \code{c(wartoscOczekiwana, odchylenieStandardowe)} rozkładów \emph{a priori}
parametrów zadań 3PL, używanych przez \emph{uirt} w estymacji; lista może zawierać
elementy o nazwach \code{a}, \code{b} lub \code{c} - najlepiej wszystkie z nich; jeśli
\emph{uirt} ma nie używać rozkładów a priori dla zadań 3PL, należy jako wartość
tego argumentu podać \code{NULL} (ale co do zasady nie jest to dobre rozwiązanie);
co do zasady domyślną wartość tego argumentu należy uznać za rozsądną}
}
\value{
lista z wynikami skalowania zawierająca następujące elementy:
\emph{grupy} (ramka danych), \emph{parametry} (ramka danych), \emph{kowariancje}
(ramka danych), \emph{usunieteZadania} (wektor ciągów znaków, może mieć zerową
długość)
}
\description{
Skaluje wyniki egzaminu uruchamiając funkcje pakietu \emph{uirt} w Stacie.
}
\seealso{
\code{\link[=obrob_wyniki_uirt]{obrob_wyniki_uirt()}}, \code{\link[=skaluj_az_dobrze]{skaluj_az_dobrze()}}
}
