% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unormuj_parametry_egzaminow.R
\name{unormuj_parametry_egzaminow}
\alias{unormuj_parametry_egzaminow}
\title{Przygotowywanie danych do obliczenia wskaznikow EWD}
\usage{
unormuj_parametry_egzaminow(parametry, dane)
}
\arguments{
\item{parametry}{lista ramek danych zwrócona przez
\code{\link[=pobierz_parametry_egzaminow]{pobierz_parametry_egzaminow()}}}

\item{dane}{lista ramek danych zwrócona przez \code{\link[=przygotuj_dane_do_ewd_bk]{przygotuj_dane_do_ewd_bk()}}}
}
\value{
lista czterech ramek danych z elementami \emph{parametryMatura},
\emph{kowariancjeMatura}, \emph{parametryWejscie} i \emph{kowariancjeWejscie}
}
\description{
Unormowuje parametry egzaminów w taki sposób, aby w populacji objętych
obliczaniem wskaźników EWD rozkład umiejętności zarówno na wejściu jak i na
wyjściu miał wartość oczekiwaną 0 i odchylenie standardowe 1.
}
\seealso{
\code{\link[=unormuj_parametry_egzaminu]{unormuj_parametry_egzaminu()}}, \code{\link[=oblicz_ewd_bk]{oblicz_ewd_bk()}}
}
