% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unormuj_parametry_egzaminow.R
\name{unormuj_parametry_egzaminu}
\alias{unormuj_parametry_egzaminu}
\title{Przygotowywanie danych do obliczenia wskaznikow EWD}
\usage{
unormuj_parametry_egzaminu(parametry, kowariancje, dane)
}
\arguments{
\item{parametry}{ramka danych z wartościami parametrów modelu skalowania
egzaminu}

\item{kowariancje}{ramka danych z kowariancjami parametrów modelu skalowania
egzaminu}

\item{dane}{ramka danych z wynikami egzaminu (używana do obliczenia rozkładu
częstości grup w populacji objętych obliczaniem wskaźników EWD)}
}
\value{
lista dwóch ramek danych z elementami \emph{parametry} i \emph{kowariancje}
}
\description{
\emph{Koń roboczy} \link{unormuj_parametry_egzaminow}. Unormowuje parametry danego
egzaminu w taki sposób, aby w populacji objętych obliczaniem wskaźników EWD
rozkład umiejętności  miał wartość oczekiwaną 0 i odchylenie standardowe 1.
}
