% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usun_wyniki_laureatow.R
\name{usun_wyniki_laureatow}
\alias{usun_wyniki_laureatow}
\title{Przetwarzanie danych z wynikami egzaminow}
\usage{
usun_wyniki_laureatow(
  dane,
  czesciEgzaminow,
  tylkoRozprawki = FALSE,
  usunPusteWiersze = TRUE
)
}
\arguments{
\item{dane}{ramka danych z wynikami egzaminu}

\item{czesciEgzaminow}{ramka danych z charakterystykami (pseudo)kryteriów
oceny skali (występujących w wynikach przekazanych argumentem \code{dane}),
zwrócona przez [przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())]

[przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())]: R:przygotuj_mapowanie_kryteriow_na_czesci_egzaminu())}

\item{tylkoRozprawki}{wartość logiczna - czy usunąć tylko wyniki
(pseud)kryteriów oceny związanych z rozprawką (wypracowaniem)?}

\item{usunPusteWiersze}{wartość logiczna - czy usunąć z \code{dane} wiersze,
które po usunięciu wyników zawierają same braki danych?}
}
\value{
ramka danych przekazana argumentem \code{dane}, z niektórymi wartościami
zamienionymi na braki danych i ew., jeśli \code{usunPusteWiersze=TRUE}, usuniętymi
niektórymi wierszami
}
\description{
Funkcja pozwala usunąć z danych wyniki laureatów - wszystkie albo tylko
w (pseudo)kryteriach oceny odnoszących się do rozprawki (wypracowania)
}
\details{
(Pseudo)kryteria oceny rozprawek (wypracowań) są w kontekście laureatów
szczególnie problematyczne, gdyż dla laureatów wybór tematu jest
nieokreślony.
}
\seealso{
\code{\link[=przygotuj_dane_do_skalowania]{przygotuj_dane_do_skalowania()}}
}
