% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_pliki_z_wartosciami_ewd.R
\name{wczytaj_plik_z_wartosciami_ewd}
\alias{wczytaj_plik_z_wartosciami_ewd}
\title{Obliczanie "wieloletnich" wskaznikow EWD poprzez agregacje wskaznikow "jednorocznych"}
\usage{
wczytaj_plik_z_wartosciami_ewd(plik)
}
\arguments{
\item{plik}{ciągów znaków z nazwą (ścieżką do) pliku}
}
\value{
lista z elementami \emph{ewd} (lista ramek danych), \emph{wskazniki},
\emph{wskaznikiSkalowania} i \emph{liczbaZdajacych} (ramki danych)
}
\description{
Wczytuje pojedynczy plik \emph{.RData} z wartościami wskaźników EWD i kontroluje
spójność jego zawartości.
}
\details{
W stosunku do formatu, w jakim dane przechowywane są w wejściowych plikach,
do elementu \emph{liczbaZdajacych} dopisywana jest kolumna \code{rok_do}.
}
\seealso{
\code{\link[=wczytaj_pliki_z_wartosciami_ewd]{wczytaj_pliki_z_wartosciami_ewd()}}
}
