% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_pliki_z_wartosciami_ewd.R
\name{wczytaj_pliki_z_wartosciami_ewd}
\alias{wczytaj_pliki_z_wartosciami_ewd}
\title{Obliczanie "wieloletnich" wskaznikow EWD poprzez agregacje wskaznikow "jednorocznych"}
\usage{
wczytaj_pliki_z_wartosciami_ewd(pliki, wskazniki, lata, wskaznikiWR)
}
\arguments{
\item{pliki}{wektor ciągów znaków z nazwami (ścieżkami do) plików}

\item{wskazniki}{wektor ciąg znaków z nazwami wskaźników, które mają zostać
zagregowane lub wyrażeniami regularnymi, do których nazwy takich wskaźników
mają pasować - por. argument \code{wskaznikiWR}}

\item{lata}{wektor liczb naturalnych - lata, które mają zostać
zagregowane}

\item{wskaznikiWR}{wartość logiczna - czy traktować argument \code{wskazniki}jako
zawierający wyrażenia regularne (a nie nazwy wskaźników)? domyślnie \code{TRUE},
jeśli argument \code{wskaźniki} to pojedynczy ciąg znaków, a \code{FALSE}, jeśli jest
to wektor kilku ciągów znaków}
}
\value{
lista ramek danych z elementami \emph{ewd}, \emph{wskazniki},
\emph{wskaznikiSkalowania} i \emph{liczbaZdajacych}
}
\description{
Wczytuje pliki \emph{.RData} z wartościami wskaźników EWD.
}
\seealso{
\code{\link[=wczytaj_plik_z_wartosciami_ewd]{wczytaj_plik_z_wartosciami_ewd()}}, \code{\link[=agreguj_ewd]{agreguj_ewd()}}
}
