% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_wyniki_surowe.R
\name{wczytaj_wyniki_surowe}
\alias{wczytaj_wyniki_surowe}
\title{Wczytywanie wynikow egzaminow zapisanych na dysku}
\usage{
wczytaj_wyniki_surowe(
  katalogDane,
  rodzajEgzaminu,
  rok,
  idSkali,
  kryteria = NULL,
  usunObserwacjeBezWynikow = TRUE,
  src = NULL
)
}
\arguments{
\item{katalogDane}{ciąg znaków - ścieżka do katalogu, w którym znajdują
się dane z wynikami surowymi egzaminów, pobranymi przy pomocy funkcji
\code{\link[EWDdane]{pobierz_wyniki_surowe}}}

\item{rodzajEgzaminu}{ciąg znaków}

\item{rok}{liczba naturalna}

\item{idSkali}{liczba naturalna - id_skali w bazie dla skali, która ma zostać
zastosowana do wyników surowych}

\item{kryteria}{opcjonalnie wektor tekstowy z nazwami wszystkich
kryteriów oceny, jakie powinna zawierać dana część/ci egzaminu}

\item{usunObserwacjeBezWynikow}{wartość logiczna wskazująca, czy
z przygotowanego zbioru powinny zostać usunięte obserwacje, które nie mają
żadnych wyników (niebędących brakami danych) kryteriów oceny przypisanej do
danej skali (i ew. zawężonych do podanych argumentem `kryteria`) - domyślnie
prawda (w praktyce ma zastosowanie w 2023 r. w związku z przeprowadzeniem
w jednym roku matur w starszej i nowszej formule)}

\item{src}{NULL połączenie z bazą danych IBE zwracane przez funkcję
\code{\link[ZPD]{polacz}}. Jeśli nie podane, podjęta zostanie próba
automatycznego nawiązania połączenia.}
}
\value{
data frame (data table)
}
\description{
Funkcja wczytuje wyniki surowe egzaminu, zapisane wcześniej na dysku
funkcją \code{\link[EWDdane]{pobierz_wyniki_surowe}} z pakietu EWDdane.
}
\details{
Funkcja co do zasady dołącza do wyników egzaminu dane kontekstowe, zawężając
grupę zwracanych obserwacji do tych, dla których te dane istnieją. Jeśli
jednak stwierdzi, że w efekcie usunięci zostaliby wszyscy, to zwróci wyniki,
bez dołączania do nich danych kontekstowych. Dojdzie do tego w szczególności
przy wczytywaniu wyników egzaminu gimnazjalnego z lat wcześniejszych niż
2006 r., gdyż są to dane CKE i dane pobierane funkcją
\code{\link[EWDdane]{pobierz_dane_kontekstowe}} ich nie obejmują.
}
