% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyswietl_rozklad_grup.R
\name{wyswietl_rozklad_grup}
\alias{wyswietl_rozklad_grup}
\title{Przetwarzanie danych z wynikami egzaminow}
\usage{
wyswietl_rozklad_grup(dane)
}
\arguments{
\item{dane}{ramka danych z wynikami egzaminu - musi zawierać kolumnę \emph{grupa}}
}
\value{
ramka danych z rozkładem
}
\description{
Drukuje na konsolę rozkład przypisania uczniów do grup wyróżnianych
w procesie skalowania.
}
\seealso{
\code{\link[=okresl_grupe]{okresl_grupe()}}, \code{\link[=przygotuj_dane_do_ewd_bk]{przygotuj_dane_do_ewd_bk()}}
}
