% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapisz_dane_dla_pvreg.R
\name{zapisz_dane_dla_pvreg}
\alias{zapisz_dane_dla_pvreg}
\title{Obliczanie latentnych wskaznikow EWD}
\usage{
zapisz_dane_dla_pvreg(dane, parametry, nazwa, nPV = 3L, nWatkow = 3L)
}
\arguments{
\item{dane}{lista ramek danych zwrócona przez \code{\link[=przygotuj_dane_do_ewd_bk]{przygotuj_dane_do_ewd_bk()}}}

\item{parametry}{lista ramek danych zwrócona przez
\code{\link[=pobierz_parametry_egzaminow]{pobierz_parametry_egzaminow()}}, a następnie zmodyfikowana przez
\code{\link[=unormuj_parametry_egzaminow]{unormuj_parametry_egzaminow()}}}

\item{nazwa}{ciąg znaków - podstawa do tworzenia nazw plików z danymi dla
\emph{pvreg} i wynikami zapisywanymi przez \emph{pvreg}}

\item{nPV}{liczba (naturalna) - liczba PV z indywidualnymi oszacowaniami
umiejętności, które mają zostać wykorzystane przez \emph{pvreg}}

\item{nWatkow}{liczba (naturalna) - liczba równoległych wątków, w których
\emph{pvreg} będzie prowadził estymację (co do zasady powinna być dzielnikiem
\code{nPV})}
}
\value{
ciąg znaków z nazwą pliku sterującego (w formacie JSON) z parametrami
wywołania \emph{pvreg}
}
\description{
Zapisuje pliki z danymi konieczne do obliczenia wskaźników EWD przez pakiet
\emph{pvreg} Pythona.
}
\seealso{
\code{\link[=przygotuj_zbior_z_parametrami_dla_pvreg]{przygotuj_zbior_z_parametrami_dla_pvreg()}}, \code{\link[=estymuj_pvreg]{estymuj_pvreg()}}
}
