% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/znajdz_skale.R
\name{znajdz_skale}
\alias{znajdz_skale}
\title{Wyszukiwanie skal}
\usage{
znajdz_skale(skale, doPrezentacji = NA, skalowanie = -1L, src = NULL)
}
\arguments{
\item{skale}{wektor ciągów znaków z wyrażeniami regularnymi
identyfikującymi skale (po kolumnie \code{opis} w bazie danych) lub wektor
liczbowy z \code{id_skali} - wskazuje skale, które mają zostać wykorzystane do
skalowania; domyślna wartość powinna działać poprawnie (tworząc wyrażenie
regularne na podstawie wartości argumentu \code{rokEWD}) w latach nie
zahaczających (zakresem danych) o zmiany formuły egzaminów}

\item{doPrezentacji}{wartość logiczna (domyślnie \verb{NA``) - czy szukać tylko skal oznaczonych w bazie danych jako przeznaczone do prezentacji? domyślna wartość }NA` wskazuje, aby szukać wszystkich skal, bez względu na to, czy
są one oznaczone jako przeznaczone do prezentacji, czy nie}

\item{skalowanie}{numer skalowania (w ramach każdej ze skal
zidentyfikowanych na podstawie argumentu \code{skale}), parametry z którego
mają zostać wykorzystane do skalowania wyników egzaminu;
domyślna wartość -1 oznacza wybór (oddzielnie w ramach każdej skali)
największego numeru skalowania spośród tych, które nie opisują wyników
działania \emph{pvreg} (tzn. pochodzą ze skalowania wyników egzaminu \emph{uirt}-em,
a nie powstały w wyniku obliczania latentnych wskaźników EWD)}

\item{src}{połączenie z bazą danych IBE zwracane przez funkcję \code{\link[ZPD:polacz]{ZPD::polacz()}};
jeśli nie podane, podjęta zostanie próba automatycznego nawiązania połączenia
(poprzez wywoływanie funkcji \code{\link[ZPD:polacz]{ZPD::polacz()}} z domyślnymi argumentami)}
}
\value{
ramka danych zawierająca id i opisy skal oraz informacje
o największym numerze skalowania przypisanym do skali i największym numerze
skalowania nie będącym skalowaniem związanym z estymacją EWD przy pomocy
\emph{pvreg} przypisanym do bazy
}
\description{
Funkcja wyszukuje skale na podstawie porównując ich opisy do podanego
wyrażenia regularnego (lub znajduje skale o podanych \emph{id_skali}) i zwraca
informacje o największych numerach skalowania powiązanych z wyszukanymi
skalami.
}
\seealso{
\code{\link[=skaluj_matura_bk]{skaluj_matura_bk()}}, \code{\link[=skaluj_we_bk]{skaluj_we_bk()}}, \code{\link[=oblicz_ewd_bk]{oblicz_ewd_bk()}}
}
