% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_az_dobrze.R
\name{zwroc_zadania_niespelniajace_kryteriow}
\alias{zwroc_zadania_niespelniajace_kryteriow}
\title{Skalowanie wynikow egzaminow w Stacie przy pomocy pakietu uirt}
\usage{
zwroc_zadania_niespelniajace_kryteriow(
  parametry,
  absA = 0.1,
  absB = 10,
  bsNotNaN = FALSE
)
}
\arguments{
\item{parametry}{ramka danych z wyestymowanymi wartościami parametrów zadań}

\item{absA}{liczba nieujemna - minimalna wartość bezwzględna dyskryminacji}

\item{absB}{liczba nieujemna - maksymalna wartość bezwzględna trudności}

\item{bsNotNaN}{wartość logiczna - czy jako niespełniające kryteriów
traktować także zadania, dla których nie udało się wyestymować błędu
standardowego któregoś z parametrów?}
}
\value{
ciąg znaków z nazwami zadań
}
\description{
Zwraca nazwy zadań, których wyestymowane parametry nie spełniają podanych
kryteriów.
}
\seealso{
\code{\link[=skaluj_az_dobrze]{skaluj_az_dobrze()}}
}
