% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_matura_bk.R
\name{skaluj_matura_bk}
\alias{skaluj_matura_bk}
\title{Obliczanie latentnych wskaznikow EWD szkol maturalnych}
\usage{
skaluj_matura_bk(
  rokEWD,
  skale = paste0("^ewd;m_[^;]+;", rokEWD),
  skalujTylkoGdyBrakSkalowan = TRUE,
  zapisz = TRUE,
  katalogSurowe = "../../dane surowe",
  nadpiszWynikiSkalowaniaNaDysku = FALSE,
  maxNIter = 1000L,
  uzyjRozkladowAPrioriDlaZadan3PL = list(a = c(1.5, 2), b = c(0, 2.5), c = c(2.5, 5.5)),
  kryteriaUsuwaniaZadan = list(absA = 0.1, absB = 10),
  src = NULL
)
}
\description{
Obliczanie latentnych wskaznikow EWD szkol maturalnych
}
